Get Clouds from distance matrix

Read distance matrices from different models, run dimensional reduction for visualization based on different techniques and store the coordinates corresponding to each model in a dataframe per technique. The names of the models will be found in the models file and their paths will be searched for in input_dir: if a file is not found, a warning will be issued.

Usage

getClouds(
  input_dir,
  output_dir,
  files_list,
  lemma,
  solutions,
  logrank = TRUE,
  logdist = TRUE,
  type = "token",
  row_selection = vector()
)

Arguments

input_dir

Directory where the token distance matrices are stored.

output_dir

Directory where the data will be stored.

files_list

Liste of filenames within input_dir.

lemma

Name of the lemma, for filenames

solutions

Named list of techniques to run for visualization possible technique values in getFit.

logrank

Whether to transform the matrices with transformMats.

logdist

Whether euclidean distances should be computed between the rows of the transformed matrices (when logrank is TRUE). Otherwise, the matrix of log-transformed ranks is only made symmetric.

type

Whether to open the files with tokensFromPac (for "token") or focdistsFromCsv (otherwise).

row_selection

List of row (and column) names to subset the matrices.

Value

List of stresses (emtpy if "mds" is not given.)